pdf-scan/requirements.txt
pymupdf
pytesseract
pdf2image
Pillow
rapidfuzz
pandas
typer
